added cart functionality

+++ routes/api/cart.js
@@ -27,19 +27,37 @@ route.get('/', (req,res)=> {
     
 })
 
-route.post('/delete',(req,res)=>{
-    Cart.destroy({
-        where:{productid:req.body.id}
+route.post('/delete',async (req,res)=>{
+
+
+    var exist= await Cart.findOne({
+        where:{productid: req.body.id}
     })
-    .then(()=>{
-        console.log("after cart deletion")
-        res.send(Cart.findAll())
-    })
-    .catch((err) => {
-    res.status(500).send({
-        error: "Could not add to carts"
-    })
-    })
+    
+    // console.log("exist object is....",exist)
+    if(exist.qty==1){
+        Cart.destroy({
+            where:{productid:req.body.id}
+        })
+        .then(()=>{
+            console.log("after cart deletion")
+            res.send(Cart.findAll())
+        })
+        .catch((err) => {
+        res.status(500).send({
+            error: "Could not add to carts"
+        })
+        })
+    }
+    else{
+        console.log("does exist changed qty.......")
+        Cart.update(
+         {qty:exist.qty-1},
+        { where:{productid:req.body.id}}
+        )
+    }
+
+    
 })
 
 route.post('/', async (req, res) => {
@@ -54,29 +72,46 @@ route.post('/', async (req, res) => {
     // console.log("2",results)
     var result=JSON.parse(JSON.stringify(results))
     var v=JSON.parse(JSON.stringify(req.user))
-    console.log(result)
-    console.log(result.id)
-    console.log(result.name)
-    console.log(result.manufacturer)
-    console.log(result.price)
-    console.log(result["id"])
+    // console.log(result)
+    // console.log(result.id)
+    // console.log(result.name)
+    // console.log(result.manufacturer)
+    // console.log(result.price)
+    // console.log(result["id"])
+
+    var exist= await Cart.findOne({
+        where:{productid:result.id}
+    })
     
-    Cart.create({
-                userid:v.id,
-                productid: result.id,
-                name:result.name,
-                manufacturer:result.manufacturer,
-                price:result.price
+    // console.log("exist object is....",exist)
+    if(exist==null){
+        console.log("did not exist............")
+        Cart.create({
+            userid:v.id,
+            productid: result.id,
+            name:result.name,
+            manufacturer:result.manufacturer,
+            price:result.price,
+            qty:1
+
+        }).then((carts)=>{
+                    console.log("after cart additoion")
+                    res.status(201).send(carts)
+        })
+        .catch((err) => {
+            res.status(500).send({
+                error: "Could not add to carts"
+            })
+        })
+    }
+    else{
+        console.log("does exist changed qty.......")
+        Cart.update(
+         {qty:exist.qty+1},
+        { where:{productid:result.id}}
+        )
+    }
     
-            }).then((carts)=>{
-                        console.log("after cart additoion")
-                        res.status(201).send(carts)
-            })
-            .catch((err) => {
-                res.status(500).send({
-                    error: "Could not add to carts"
-                })
-            })
    
 })
